refactor(arch): move pulse ledger to tools/ to honor ADR-007
CI arch-check flagged the new code: src/state and src/tools cannot
import from runtime::context*, but the prior commit put LocatedRead +
pulse-ledger functions in runtime/context_cache.rs and imported them
from routing.rs, send_message.rs, file_ops.rs, and the probe test.

This commit reverts runtime/context_cache.rs to its origin/main state
and lifts the ledger out to a stand-alone tools/pulse_ledger.rs that
state and tools may both depend on without crossing the layer boundary.
The cache-aware range read introduced in the prior commit is dropped
for now; the read_file routing arm goes back to tool_operator
.read_file_range and records the claimed [offset, offset + limit - 1]
range. A wider-than-actual window is harmless to the anchor logic and
keeps the layering clean. ADR-033 amendment notes the placement choice
explicitly.
