+++ src/tools.rs
@@ -1,6 +1,7 @@
 pub mod embed;
 pub mod index;
 pub(crate) mod operator;
+pub(crate) mod pulse_ledger;
 pub mod search;
 pub mod semantic;
 pub(crate) mod workspace_explore;
